+++ Reinvest/LegalEntities/src/Providers/EventBusProvider.ts
@@ -16,7 +16,7 @@ export default class EventBusProvider {
 
         eventBus
             .subscribeHandlerForKinds(SendToQueueEventHandler.getClassName(), [
-                "IndividualAccountOpened",
+                "LegalProfileCompleted",
             ]);
         ;