+++ Reinvest/Registration/migrations/index.ts
@@ -2,3 +2,4 @@ export * as migration_20230320154400_reg_create_mapping_registry from './registr
 export * as migration_20230322125000_create_north_capital_synchronization
     from './registration_2_create_north_capital_synchronization';
 export * as migration_20230322151500_update_size_of_mapped_type from './registration_3_update_size_of_mapped_type';
+export * as migration_20230323092000_add_links_mapping from './registration_4_add_links_mapping';