+++ Reinvest/Registration/src/Adapter/Database/Repository/NorthCapitalSynchronizationRepository.ts
@@ -1,7 +1,7 @@
 import {IdGeneratorInterface} from "IdGenerator/IdGenerator";
 import {
     northCapitalSynchronizationTable,
-    RegistrationDatabaseAdapterProvider,
+    RegistrationDatabaseAdapterProvider, registrationMappingRegistryTable,
 } from "Registration/Adapter/Database/DatabaseAdapter";
 import {
     NorthCapitalEntityType,
@@ -9,9 +9,11 @@ import {
     NorthCapitalSynchronizationRecordType
 } from "Registration/Adapter/NorthCapital/NorthCapitalSynchronizationRecord";
 import {
-    InsertableNorthCapitalSynchronization,
+    InsertableNorthCapitalSynchronization, SelectablePartyId,
     SelectableSynchronizationRecord
 } from "Registration/Adapter/Database/RegistrationSchema";
+import {MappedType} from "Registration/Domain/Model/Mapping/MappedType";
+import {NorthCapitalSynchronizationMapping} from "Registration/Domain/VendorModel/NorthCapital/NorthCapitalTypes";
 
 export class NorthCapitalSynchronizationRepository {
     public static getClassName = (): string => "NorthCapitalSynchronizationRepository";
@@ -24,27 +26,32 @@ export class NorthCapitalSynchronizationRepository {
         this.idGenerator = uniqueGenerator;
     }
 
-    async getSynchronizationRecord(recordId: string): Promise<NorthCapitalSynchronizationRecord | never> {
-        const data = await this.databaseAdapterProvider.provide()
-            .selectFrom(northCapitalSynchronizationTable)
-            .select(['recordId', 'northCapitalId', 'type', 'crc', 'documents', 'version'])
-            .where('recordId', '=', recordId)
-            .limit(1)
-            .executeTakeFirstOrThrow() as SelectableSynchronizationRecord as NorthCapitalSynchronizationRecordType;
+    async getSynchronizationRecord(recordId: string): Promise<NorthCapitalSynchronizationRecord | null> {
+        try {
+            const data = await this.databaseAdapterProvider.provide()
+                .selectFrom(northCapitalSynchronizationTable)
+                .select(['recordId', 'northCapitalId', 'type', 'crc', 'documents', 'version', 'links'])
+                .where('recordId', '=', recordId)
+                .limit(1)
+                .executeTakeFirstOrThrow() as SelectableSynchronizationRecord as NorthCapitalSynchronizationRecordType;
 
-        return NorthCapitalSynchronizationRecord.create(data);
+            return NorthCapitalSynchronizationRecord.create(data);
+        } catch (error: any) {
+            return null;
+        }
     }
 
-    async createSynchronizationRecord(recordId: string, partyId: string, crc: string, entityType: NorthCapitalEntityType): Promise<void> {
+    async createSynchronizationRecord(recordId: string, northCapitalId: string, crc: string, entityType: NorthCapitalEntityType): Promise<void> {
         await this.databaseAdapterProvider.provide()
             .insertInto(northCapitalSynchronizationTable)
             .values(<InsertableNorthCapitalSynchronization>{
                 recordId,
-                northCapitalId: partyId,
+                northCapitalId,
                 crc,
                 type: entityType,
                 documents: JSON.stringify([]),
                 version: 0,
+                links: JSON.stringify([]),
             })
             .onConflict((oc) => oc
                 .columns(['recordId'])
@@ -53,7 +60,10 @@ export class NorthCapitalSynchronizationRepository {
             .execute()
     }
 
-    async updateSynchronizationRecord(synchronizationRecord: NorthCapitalSynchronizationRecord, crc: string): Promise<void> {
+    async updateSynchronizationRecord(synchronizationRecord: NorthCapitalSynchronizationRecord): Promise<void> {
+        if (!synchronizationRecord.wasUpdated()) {
+            return;
+        }
         const updatedDate = new Date();
         const nextVersion = synchronizationRecord.getNextVersion();
         const currentVersion = synchronizationRecord.getVersion();
@@ -63,8 +73,9 @@ export class NorthCapitalSynchronizationRepository {
                 .updateTable(northCapitalSynchronizationTable)
                 .set({
                     version: nextVersion,
-                    crc,
-                    updatedDate
+                    crc: synchronizationRecord.getCrc(),
+                    updatedDate,
+                    links: JSON.stringify(synchronizationRecord.getLinks()),
                 })
                 .where('recordId', '=', synchronizationRecord.getRecordId())
                 .where('version', '=', currentVersion)
@@ -73,5 +84,52 @@ export class NorthCapitalSynchronizationRepository {
         } catch (error: any) {
             console.warn(`North Capital Synchronization Record race condition detected for record ${synchronizationRecord.getRecordId()}`);
         }
+    }
+
+    async getMainPartyIdByProfile(profileId: string): Promise<string | never> {
+        try {
+            const data = await this.databaseAdapterProvider.provide()
+                .selectFrom(registrationMappingRegistryTable)
+                .fullJoin(northCapitalSynchronizationTable, `${northCapitalSynchronizationTable}.recordId`, `${registrationMappingRegistryTable}.recordId`)
+                .select([`${northCapitalSynchronizationTable}.northCapitalId`])
+                .where(`${registrationMappingRegistryTable}.profileId`, '=', profileId)
+                .where(`${registrationMappingRegistryTable}.externalId`, '=', profileId)
+                .where(`${registrationMappingRegistryTable}.mappedType`, '=', MappedType.PROFILE)
+                .limit(1)
+                .castTo<SelectablePartyId>()
+                .executeTakeFirstOrThrow();
+
+            return data.northCapitalId;
+        } catch (error: any) {
+            throw new Error('MAIN_PARTY_NOT_FOUND');
+        }
+    }
+
+    async getAllProfileSynchronizationMapping(recordId: string): Promise<NorthCapitalSynchronizationMapping[]> {
+        const {profileId} = await this.databaseAdapterProvider.provide()
+            .selectFrom(registrationMappingRegistryTable)
+            .select(['profileId'])
+            .where('recordId', '=', recordId)
+            .limit(1)
+            .executeTakeFirstOrThrow();
+
+        const data = await this.databaseAdapterProvider.provide()
+            .selectFrom(registrationMappingRegistryTable)
+            .fullJoin(northCapitalSynchronizationTable, `${northCapitalSynchronizationTable}.recordId`, `${registrationMappingRegistryTable}.recordId`)
+            .select([`${northCapitalSynchronizationTable}.northCapitalId`])
+            .select([`${registrationMappingRegistryTable}.profileId`, `${registrationMappingRegistryTable}.externalId`, `${registrationMappingRegistryTable}.mappedType`])
+            .where(`${registrationMappingRegistryTable}.profileId`, '=', profileId)
+            .execute();
+
+        return data.map((row: any) => {
+            return {
+                mapping: {
+                    type: row.mappedType,
+                    profileId: row.profileId,
+                    externalId: row.externalId,
+                },
+                northCapitalId: row.northCapitalId,
+            }
+        });
     }
 }