+++ Reinvest/Registration/src/Domain/VendorModel/NorthCapital/IndividualAccount.ts
@@ -1,16 +1,24 @@
 import {CrcService} from "Registration/Domain/CrcService";
-import {IndividualAccountType} from "Registration/Domain/VendorModel/NorthCapital/NorthCapitalTypes";
+import {
+    IndividualAccountType,
+    IndividualExtendedMainPartyType,
+    NorthCapitalIndividualAccountType, NorthCapitalLink
+} from "Registration/Domain/VendorModel/NorthCapital/NorthCapitalTypes";
 
 import {IndividualAccountForSynchronization} from "Registration/Domain/Model/Account";
 import {NorthCapitalMapper} from "Registration/Domain/VendorModel/NorthCapital/NorthCapitalMapper";
+import {MappedType} from "Registration/Domain/Model/Mapping/MappedType";
 
 export class IndividualAccount {
     private readonly data: IndividualAccountType;
-    private readonly crc: string;
+    private accountCrc: string;
+    private partyCrc: string;
+
 
     constructor(data: IndividualAccountType) {
         this.data = data;
-        this.crc = this.generateCrc(data);
+        this.accountCrc = this.generateAccountCrc(data.account);
+        this.partyCrc = this.generateParty(data.extendedParty);
     }
 
     static createFromIndividualAccountForSynchronization(data: IndividualAccountForSynchronization): IndividualAccount | never {
@@ -34,7 +42,6 @@ export class IndividualAccount {
             account: {
                 accountRegistration: fullName,
                 type: "Individual",
-                domesticYN: "domestic_account",
                 streetAddress1: data.address.addressLine1,
                 streetAddress2: data.address?.addressLine2,
                 city: data.address.city,
@@ -42,34 +49,77 @@ export class IndividualAccount {
                 zip: data.address.zip,
                 country: data.address.country,
             },
+            links: [
+                {
+                    mappingConfiguration: {
+                        type: MappedType.PROFILE,
+                        profileId: data.profileId,
+                        externalId: data.profileId,
+                    },
+                    linkConfiguration: {
+                        relatedEntryType: "IndivACParty",
+                        linkType: "owner",
+                        primary_value: 1
+                    }
+                }
+            ]
         });
     }
 
-    getData(): IndividualAccountType {
-        return this.data;
+    private generateAccountCrc(data: IndividualAccountType['account']): string {
+        const values = [
+            data.accountRegistration,
+            data.streetAddress1,
+            data.streetAddress2 ?? "",
+            data.city,
+            data.state,
+            data.zip,
+            data.country,
+        ];
+
+        return CrcService.generateCrc(values);
     }
 
-    private generateCrc(data: IndividualAccountType): string {
+    private generateParty(data: IndividualAccountType['extendedParty']): string {
         const values = [
-            data.profileId,
-            data.extendedParty.occupation ?? "",
-            data.extendedParty.empStatus ?? "",
-            data.extendedParty.empName ?? "",
-            data.extendedParty.householdNetworth ?? "",
-            data.extendedParty.currentHouseholdIncome ?? "",
-            data.account.accountRegistration,
-            data.account.streetAddress1,
-            data.account.streetAddress2 ?? "",
-            data.account.city,
-            data.account.state,
-            data.account.zip,
-            data.account.country,
+            data.occupation ?? "",
+            data.empStatus ?? "",
+            data.empName ?? "",
+            data.householdNetworth ?? "",
+            data.currentHouseholdIncome ?? "",
         ];
 
         return CrcService.generateCrc(values);
     }
 
     getCrc(): string {
-        return this.crc;
+        return JSON.stringify({
+            account: this.accountCrc,
+            party: this.partyCrc,
+        });
+    }
+
+    getProfileId(): string {
+        return this.data.profileId;
+    }
+
+    getPartyData(): IndividualExtendedMainPartyType {
+        return this.data.extendedParty;
+    }
+
+    getAccountData(): NorthCapitalIndividualAccountType {
+        return this.data.account;
+    }
+
+    getLinksConfiguration(): NorthCapitalLink[] {
+        return this.data.links;
+    }
+
+    isOutdatedAccount(crc: string): boolean {
+        return this.accountCrc !== JSON.parse(crc)?.account;
+    }
+
+    isOutdatedParty(crc: string): boolean {
+        return this.partyCrc !== JSON.parse(crc)?.party;
     }
 }