+++ Reinvest/Registration/src/Domain/VendorModel/NorthCapital/MainParty.ts
@@ -3,6 +3,7 @@ import {CrcService} from "Registration/Domain/CrcService";
 import {MainPartyType} from "Registration/Domain/VendorModel/NorthCapital/NorthCapitalTypes";
 import DateTime from "date-and-time";
 import {NorthCapitalMapper} from "Registration/Domain/VendorModel/NorthCapital/NorthCapitalMapper";
+import {DictionaryType} from "HKEKTypes/Generics";
 
 export class MainParty {
     private data: MainPartyType;
@@ -36,10 +37,6 @@ export class MainParty {
         });
     }
 
-    getData(): MainPartyType {
-        return this.data;
-    }
-
     private generateCrc(data: MainPartyType): string {
         const values = [
             data.domicile ?? "",
@@ -60,15 +57,29 @@ export class MainParty {
         return CrcService.generateCrc(values);
     }
 
-    getEmail(): string {
-        return this.data.emailAddress;
-    }
-
-    getLastName() {
-        return this.data.lastName;
-    }
-
     getCrc(): string {
         return this.crc;
+    }
+
+    getPartyData(): DictionaryType {
+        const rawData = this.data as DictionaryType;
+        const data = {} as DictionaryType
+        for (const key of Object.keys(rawData)) {
+            switch (key) {
+                case 'middleInitial':
+                case 'socialSecurityNumber':
+                    if (rawData[key] && rawData[key].length > 0) {
+                        data[key] = rawData[key];
+                    }
+                    break;
+                case 'documents':
+                    break;
+                default:
+                    data[key] = rawData[key];
+                    break;
+            }
+        }
+
+        return data;
     }
 }